refactor: rename movie list related classes and interfaces
