+++ data/src/main/java/com/london/data/remote/service/list/CustomMovieListsApiService.kt
@@ -13,7 +13,7 @@ import retrofit2.http.POST
 import retrofit2.http.Path
 import retrofit2.http.Query
 
-interface MovieCustomListsApiService {
+interface CustomMovieListsApiService {
 
     @POST("3/list")
     suspend fun create(
