+++ data/src/main/java/com/london/data/remote/source/list/CustomMovieListsRemoteDataSource.kt
@@ -3,17 +3,15 @@ package com.london.data.remote.source.list
 import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.list.CreateCustomListResponse
 import com.london.data.remote.model.list.CustomListResponse
+import com.london.data.remote.model.list.CustomMovieListResponse
 import com.london.data.remote.model.search.MovieRemote
 
-interface MovieCustomListsRemoteDataSource {
+interface CustomMovieListsRemoteDataSource {
 
     suspend fun create(name: String): Result<CreateCustomListResponse>
-
     suspend fun delete(listId: Int): Result<CustomListResponse>
-
     suspend fun getDetails(listId: Int, page: Int): Result<ApiResponse<MovieRemote>>
-
     suspend fun addMovieToList(listId: Int, movieId: Int): Result<CustomListResponse>
-
+    suspend fun getAllMovieLists(page: Int): Result<ApiResponse<CustomMovieListResponse>>
     suspend fun removeMovieFromList(listId: Int, movieId: Int): Result<CustomListResponse>
 }
